chore: extract connector's related types to shared package and re-use them in both sdk and connectors
