+++ packages/connector-types/package.json
@@ -1,0 +1,18 @@
+{
+    "name": "@chili-studio/connector-types",
+    "version": "1.0.0",
+    "private": true,
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "files": [
+        "dist"
+    ],
+    "scripts": {
+        "build": "tsc",
+        "format": "prettier --write ."
+    },
+    "devDependencies": {
+        "prettier": "^3.0.3",
+        "typescript": "^5.0.0"
+    }
+}
