+++ packages/connectors/package.json
@@ -2,16 +2,20 @@
     "name": "@chili-publish/studio-connectors",
     "version": "1.23.1-rc.3",
     "types": "dist/index.d.ts",
+    "type": "module",
     "files": [
         "dist"
     ],
     "license": "MIT",
     "devDependencies": {
         "prettier": "^3.0.3",
-        "typescript": "^5.2.2"
+        "typescript": "^5.2.2",
+        "vite": "^6.3.5",
+        "vite-plugin-dts": "^4.5.3",
+        "@types/node": "^22.15.18"
     },
     "scripts": {
-        "build": "tsc",
+        "build": "vite build",
         "prepare-upload": "yarn build && yarn pack && mkdir ./upload/connectors/ -p && cp *.tgz ./upload/connectors/"
     }
 }
