+++ packages/connectors/src/MediaConnector.ts
@@ -1,4 +1,14 @@
-import { ArrayBufferPointer, ConnectorConfigValue, Dictionary, QueryOptions } from './Connector.Shared';
+import {
+    ConnectorConfigOptions,
+    Dictionary,
+    MediaConnectorCapabilities,
+    MediaDetail,
+    MediaPage,
+    QueryOptions,
+} from '@chili-studio/connector-types';
+import { ArrayBufferPointer } from './Connector.Shared';
+
+export type { Media, MediaConnectorCapabilities, MediaDetail, MediaPage } from '@chili-studio/connector-types';
 
 export interface MediaConnector {
     detail(id: string, context: Dictionary): Promise<MediaDetail>;
@@ -9,38 +19,9 @@ export interface MediaConnector {
         intent: DownloadIntent,
         context: Dictionary,
     ): Promise<ArrayBufferPointer>;
-    getConfigurationOptions(): ConnectorConfigValue[] | null;
+    getConfigurationOptions(): ConnectorConfigOptions | null;
     getCapabilities(): MediaConnectorCapabilities;
 }
 
 export type DownloadIntent = 'web' | 'print' | 'animation';
 export type DownloadType = 'thumbnail' | 'mediumres' | 'highres' | 'fullres' | 'original';
-
-export interface MediaPage {
-    pageSize: number;
-    data: Media[];
-    links: {
-        nextPage: string;
-    };
-}
-
-export interface Media {
-    id: string;
-    name: string;
-    relativePath: string;
-    type: number;
-    metaData: Dictionary;
-    extension?: string;
-}
-
-export interface MediaDetail extends Media {
-    width?: number;
-    height?: number;
-}
-
-export type MediaConnectorCapabilities = {
-    query: boolean;
-    detail: boolean;
-    filtering: boolean;
-    metadata?: boolean;
-};
