+++ packages/connectors/tsconfig.json
@@ -1,9 +1,19 @@
 {
     "compilerOptions": {
+      "target": "ES2020",
+      "module": "ESNext",
+      "moduleResolution": "bundler",
       "declaration": true,
       "emitDeclarationOnly": true,
-      "outDir": "dist"
+      "outDir": "dist",
+      "esModuleInterop": true,
+      "strict": true,
+      "skipLibCheck": true,
+      "types": ["node"],
+      "paths": {
+        "@chili-studio/connector-types": ["../connector-types/src"]
+      }
     },
-    "include": ["src"],
+    "include": ["src", "../connector-types/src", "vite.config.ts"],
     "exclude": ["node_modules", "**/__tests__/*"]
   }
