+++ packages/sdk/src/index.ts
@@ -70,9 +70,9 @@ export {
     Day,
     Locale,
     VariableType,
+    VariableVisibilityOperator,
     VariableVisibilityTargetType,
     VariableVisibilityType,
-    VariableVisibilityOperator,
 } from './types/VariableTypes';
 
 export type {
@@ -107,18 +107,15 @@ export type {
     FrameAnimationPropertiesType,
     FrameAnimationType,
 } from './types/AnimationTypes';
-export { ConnectorConfigValueType, DataRowAsyncError, ActionAsyncError } from './types/CommonTypes';
+export { ActionAsyncError, DataRowAsyncError } from './types/CommonTypes';
 export type {
     ConfigType,
-    ConnectorConfigOptions,
-    ConnectorConfigValue,
     EditorResponse,
     Id,
     InitialStateType,
-    MetaData,
     PageType,
-    SelectedLayoutFrame,
     PrivateData,
+    SelectedLayoutFrame,
 } from './types/CommonTypes';
 export type {
     ActionCircularDocumentIssue,
