+++ packages/sdk/src/types/CommonTypes.ts
@@ -9,6 +9,7 @@ import { DocumentColor } from './ColorStyleTypes';
 import { StudioOptionsDeltaUpdate, StudioStyling } from './ConfigurationTypes';
 import { AuthCredentials, AuthRefreshRequest, ConnectorEvent } from './ConnectorTypes';
 import { DocumentIssue, DocumentType, UndoState } from './DocumentTypes';
+import { EngineEditMode } from './EngineEditModeTypes';
 import type { FrameType } from './FrameTypes';
 import { Frame, FrameLayoutType, FrameTypeEnum } from './FrameTypes';
 import { LayoutListItemType, LayoutPropertiesType, LayoutWithFrameProperties, MeasurementUnit } from './LayoutTypes';
@@ -18,7 +19,6 @@ import { CornerRadiusUpdateModel } from './ShapeTypes';
 import { SelectedTextStyle } from './TextStyleTypes';
 import { Variable } from './VariableTypes';
 import { Viewport } from './ViewportTypes';
-import { EngineEditMode } from './EngineEditModeTypes';
 
 export type Id = string;
 export type BaseConfigType = {
@@ -394,28 +394,6 @@ export interface SelectedLayoutFrame {
     type: FrameTypeEnum;
     isVisible: boolean;
 }
-
-export interface MetaData {
-    [key: string]: string | boolean;
-}
-
-export interface ConnectorOptions {
-    [key: string]: string | boolean;
-}
-
-export enum ConnectorConfigValueType {
-    text = 'text',
-    boolean = 'boolean',
-}
-
-export interface ConnectorConfigValue {
-    readonly name: string;
-    readonly displayName: string;
-    readonly type: ConnectorConfigValueType;
-}
-
-export type ConnectorConfigOptions = ConnectorConfigValue[];
-
 export interface ActionEventErrorData {
     event: ActionEditorEvent;
     actionIds: Id[];
