+++ packages/sdk/src/types/ConnectorTypes.ts
@@ -1,4 +1,7 @@
+import type { Dictionary as ContextDictionary, QueryPage as QueryResult } from '@chili-studio/connector-types';
 import { Id } from './CommonTypes';
+
+export type { ConnectorConfigOptions, ConnectorConfigValue, QueryOptions } from '@chili-studio/connector-types';
 
 export enum DeprecatedMediaType {
     file = 0,
@@ -32,14 +35,10 @@ export enum SortOrder {
     descending = 'desc',
 }
 
-export type QueryOptions = {
-    filter?: string[] | null;
-    collection?: string | null;
-    pageToken?: string | null;
-    pageSize?: number | null;
-    sortBy?: SortBy | null;
-    sortOrder?: SortOrder | null;
-};
+export enum ConnectorConfigValueType {
+    text = 'text',
+    boolean = 'boolean',
+}
 
 interface ConnectorRegistrationBase {
     /**
@@ -98,7 +97,7 @@ export enum ConnectorRegistrationSource {
 
 export interface EngineToConnectorMapping {
     name: string;
-    value: string | boolean;
+    value: ContextDictionary[keyof ContextDictionary];
     direction: ConnectorMappingDirection.engineToConnector;
 }
 
@@ -157,10 +156,8 @@ export type ConnectorEvent = {
     type: ConnectorEventType;
 };
 
-export type QueryPage<T> = {
-    pageSize: number;
+export type QueryPage<T> = Omit<QueryResult<T>, 'links'> & {
     nextPageToken?: string;
-    data: T[];
 };
 
 export enum ConnectorMappingSource {
@@ -302,3 +299,6 @@ export type AuthRefreshRequest = {
     type: AuthRefreshTypeEnum;
     headerValue: string | null;
 };
+
+export type ConnectorOptions = ContextDictionary;
+export type MetaData = ContextDictionary;
