+++ packages/sdk/src/types/DataConnectorTypes.ts
@@ -1,42 +1,12 @@
-export type DataConnectorCapabilities = {
-    filtering: boolean;
-    sorting: boolean;
-    model: boolean;
-};
-
-export type PageConfig = {
-    filters?: DataFilter[] | null;
-    sorting?: DataSorting[] | null;
-    continuationToken?: string | null;
-    limit: number;
-};
-
-export type DataSorting = {
-    property: string;
-    direction: 'asc' | 'desc';
-};
-
-export type DataFilter = {
-    property: string;
-    value: string;
-    type: 'contains' | 'exact';
-};
-
-// Generic DataItem
-export type DataPage<Item> = {
+export type EditorDataPage<Item> = {
     data: Item[];
     continuationToken?: string | null;
 };
 
-export type DataItem = {
-    [key: string]: string | number | boolean | Date | null;
-};
-
-export type DataModel = {
-    properties: DataModelProperty[];
-};
-
-type DataModelProperty = {
-    name: string;
-    type: 'number' | 'boolean' | 'singleLine' | 'multiLine' | 'date';
-};
+export type {
+    DataConnectorCapabilities,
+    DataItem,
+    DataModel,
+    DataPage,
+    PageConfig,
+} from '@chili-studio/connector-types';
