+++ packages/sdk/src/types/FontConnectorTypes.ts
@@ -1,27 +1,6 @@
-import { Id } from './CommonTypes';
-
-export type FontConnectorCapabilities = {
-    query: boolean;
-    detail: boolean;
-    preview: boolean;
-    filtering: boolean;
-};
+export type { FontConnectorCapabilities, FontFamily, FontStyle } from '@chili-studio/connector-types';
 
 export enum FontPreviewFormat {
     Square = 'square',
     Line = 'line',
 }
-
-export type FontFamily = {
-    id: Id;
-    name: string;
-    fontStylesCount: number;
-    extensions: string[];
-};
-
-export type FontStyle = {
-    id: Id;
-    name: string;
-    familyId: string;
-    familyName: string;
-};
