+++ packages/sdk/src/types/MediaConnectorTypes.ts
@@ -1,12 +1,4 @@
-import { Id } from './CommonTypes';
-import { MediaType } from './ConnectorTypes';
-
-export type MediaConnectorCapabilities = {
-    query: boolean;
-    detail: boolean;
-    filtering: boolean;
-    metadata?: boolean;
-};
+export type { Media, MediaConnectorCapabilities, MediaDetail } from '@chili-studio/connector-types';
 
 export enum MediaDownloadType {
     thumbnail = 'thumbnail',
@@ -17,20 +9,6 @@ export enum MediaDownloadType {
 
 export enum MediaDownloadIntent {
     web = 'web',
-}
-
-export type Media = {
-    id: Id;
-    name: string;
-    relativePath: string;
-    type: MediaType;
-    extension: string | null;
-    metaData: {
-        [key: string]: string;
-    };
-};
-
-export interface MediaDetail extends Media {
-    width?: number;
-    height?: number;
+    print = 'print',
+    animation = 'animation',
 }
